requirements.txt
pytest
allure-pytest
selene>=2.0.0rc4
Appium-Python-Client
python-dotenv~=1.0.1
pydantic~=1.10.9